wip: input event
The actual implementation needs two apps. One hand
ling input event, the other one in charge of rendering.
I still need to synchronize the app state (in particular the page)
I also need to find a seed that does not require std::time so it
does not cause runtime exception when targetting wasm